test message helper moved to fixtures dir
